Address Claude review feedback
- enforce the sample limit inside the shared browser converter
- export shared thresholds for localized Add Raster preflight
