+++ packages/processing/src/index.ts
@@ -273,6 +273,10 @@ export {
   isTiledGeoTiff,
   convertGeoTiffToCog,
   COG_WASM_COMPRESSIONS,
+  exceedsBrowserCogConversionLimit,
+  geoTiffSampleCount,
+  LARGE_BROWSER_COG_CONVERSION_SAMPLES,
+  MAX_BROWSER_COG_CONVERSION_SAMPLES,
   type CogWasmCompression,
   type ConvertGeoTiffToCogOptions,
   type GeoTiffInfo,
